[Refactor] Move storage info structs out of schema scanner
