+++ be/src/script/script.cpp
@@ -23,7 +23,6 @@
 #include "common/logging.h"
 #include "common/stack_util.h"
 #include "common/vlog_cntl.h"
-#include "exec/schema_scanner/schema_be_tablets_scanner.h"
 #include "fs/key_cache.h"
 #include "gen_cpp/olap_file.pb.h"
 #include "gutil/strings/substitute.h"
@@ -38,6 +37,7 @@
 #include "storage/lake/vacuum.h"
 #include "storage/manual_compaction.h"
 #include "storage/primary_key_dump.h"
+#include "storage/primitive/tablet_basic_info.h"
 #include "storage/storage_engine.h"
 #include "storage/tablet.h"
 #include "storage/tablet_manager.h"
