+++ be/src/storage/lake/tablet_manager.cpp
@@ -28,7 +28,6 @@
 #include "common/compiler_util.h"
 #include "common/config_lake_fwd.h"
 #include "common/system/master_info.h"
-#include "exec/schema_scanner/schema_be_tablets_scanner.h"
 #include "fmt/format.h"
 #include "fs/fs.h"
 #include "fs/fs_factory.h"
@@ -55,6 +54,7 @@
 #include "storage/lake/versioned_tablet.h"
 #include "storage/lake/vertical_compaction_task.h"
 #include "storage/metadata_util.h"
+#include "storage/primitive/tablet_basic_info.h"
 #include "storage/protobuf_file.h"
 #include "storage/rowset/segment.h"
 #include "storage/tablet_schema_map.h"
