+++ be/src/storage/lake/tablet_manager.h
@@ -31,11 +31,11 @@
 #include "storage/lake/txn_log.h"
 #include "storage/lake/types_fwd.h"
 #include "storage/options.h"
+#include "storage/primitive/tablet_basic_info.h"
 #include "storage/rowset/base_rowset.h"
 
 namespace starrocks {
 struct FileInfo;
-struct TabletBasicInfo;
 class Segment;
 class TabletSchemaPB;
 class TCreateTabletReq;
