+++ be/src/storage/primitive/tablet_basic_info.h
@@ -15,28 +15,32 @@
 #pragma once
 
 #include <cstdint>
-#include <vector>
+#include <string>
 
-#include "exec/schema_scanner.h"
-#include "storage/primitive/txn_info.h"
+#include "gen_cpp/Types_types.h" // for TStorageMedium::type
 
 namespace starrocks {
 
-class SchemaBeTxnsScanner : public SchemaScanner {
-public:
-    SchemaBeTxnsScanner();
-    ~SchemaBeTxnsScanner() override;
-
-    Status start(RuntimeState* state) override;
-    Status get_next(ChunkPtr* chunk, bool* eos) override;
-
-private:
-    Status fill_chunk(ChunkPtr* chunk);
-
-    int64_t _be_id{0};
-    std::vector<TxnInfo> _infos;
-    size_t _cur_idx{0};
-    static SchemaScanner::ColumnDesc _s_columns[];
+struct TabletBasicInfo {
+    int64_t table_id{0};
+    int64_t partition_id{0};
+    int64_t tablet_id{0};
+    int64_t num_version{0};
+    int64_t max_version{0};
+    int64_t min_version{0};
+    int64_t num_rowset{0};
+    int64_t num_segment{0};
+    int64_t num_row{0};
+    int64_t data_size{0};
+    int64_t index_mem{0};
+    int64_t create_time{0};
+    int32_t state{0};
+    int32_t type{0};
+    std::string data_dir;
+    int64_t shard_id{0};
+    int64_t schema_hash{0};
+    int64_t index_disk_usage{0};
+    TStorageMedium::type medium_type;
 };
 
 } // namespace starrocks
