+++ be/src/storage/tablet.cpp
@@ -51,7 +51,6 @@
 #include "common/config_storage_fwd.h"
 #include "common/tracer.h"
 #include "common/util/table_metrics.h"
-#include "exec/schema_scanner/schema_be_tablets_scanner.h"
 #include "runtime/current_thread.h"
 #include "runtime/exec_env.h"
 #include "storage/binlog_builder.h"
@@ -62,6 +61,7 @@
 #include "storage/default_compaction_policy.h"
 #include "storage/olap_common.h"
 #include "storage/olap_define.h"
+#include "storage/primitive/tablet_basic_info.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_meta_manager.h"
 #include "storage/size_tiered_compaction_policy.h"
