+++ be/src/storage/tablet_manager.cpp
@@ -46,7 +46,6 @@
 #include "base/testutil/sync_point.h"
 #include "common/config_compaction_fwd.h"
 #include "common/config_storage_fwd.h"
-#include "exec/schema_scanner/schema_be_tablets_scanner.h"
 #include "fs/fs.h"
 #include "fs/fs_util.h"
 #include "gutil/strings/substitute.h"
@@ -55,6 +54,7 @@
 #include "storage/data_dir.h"
 #include "storage/olap_common.h"
 #include "storage/persistent_index_load_executor.h"
+#include "storage/primitive/tablet_basic_info.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_writer.h"
 #include "storage/rowset/rowset_writer_context.h"
