+++ be/src/storage/tablet_updates.cpp
@@ -35,7 +35,6 @@
 #include "common/config_storage_fwd.h"
 #include "common/status.h"
 #include "common/tracer.h"
-#include "exec/schema_scanner/schema_be_tablets_scanner.h"
 #include "fs/fs_factory.h"
 #include "gen_cpp/MasterService_types.h"
 #include "gen_cpp/olap_file.pb.h"
@@ -59,6 +58,7 @@
 #include "storage/primary_key_dump.h"
 #include "storage/primitive/chunk_iterator.h"
 #include "storage/primitive/empty_iterator.h"
+#include "storage/primitive/tablet_basic_info.h"
 #include "storage/primitive/union_iterator.h"
 #include "storage/rows_mapper.h"
 #include "storage/rowset/base_rowset.h"
