+++ be/src/storage/txn_manager.cpp
@@ -47,8 +47,8 @@
 #include "common/runtime_profile.h"
 #include "common/thread/threadpool.h"
 #include "common/tracer.h"
-#include "exec/schema_scanner/schema_be_txns_scanner.h"
 #include "storage/data_dir.h"
+#include "storage/primitive/txn_info.h"
 #include "storage/rowset/rowset_meta_manager.h"
 #include "storage/storage_engine.h"
 #include "storage/storage_metrics.h"
