+++ be/test/storage/tablet_test.cpp
@@ -19,9 +19,9 @@
 #include <thread>
 #include <vector>
 
-#include "exec/schema_scanner/schema_be_tablets_scanner.h"
 #include "storage/data_dir.h"
 #include "storage/olap_common.h"
+#include "storage/primitive/tablet_basic_info.h"
 #include "storage/rowset/rowset.h"
 #include "storage/rowset/rowset_meta.h"
 #include "storage/tablet_manager.h"
